Redoing enumeration system (Part 1, not buildable) Changes the "namespace" enumeration to real enumerations. Minor other changes done too. If the diff is not clear, it's because of Windows, which changed the newline system, so I changed it back. Sorry for the inconvenience.

+++ src/regimys/gui/Overworld.hpp
@@ -58,7 +58,7 @@ public:
     bool movementLock = false;
 
     bool scrollock[2] = {false, false};
-    int ppDir = Side::TO_DOWN;
+  int ppDir = (int) Side::TO_DOWN;
 
     bool scrolling = true;
